gudangObatv4/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.3.0
Markdown==3.3.3
oauthlib==3.1.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
whitenoise==5.2.0
wincertstore==0.2
zipp==3.4.0
